+++ src/http/utils-server/fdalabel.ts
@@ -1,0 +1,5 @@
+"use server";
+import { cookies } from "next/headers";
+import { defineCookieAttr } from "./generic";
+
+export async function setupScrapeVersionsHook() {}
